pyproject.toml
[tool.pdm]
[tool.pdm.dev-dependencies]
dev = [
    "black>=23.3.0",
]

[project]
name = ""
version = ""
description = ""
authors = [
    {name = "MiguelBochi", email = "miguelbochibarro@gmail.com"},
]
dependencies = [
    "django>=4.2.2",
    "djangorestframework>=3.14.0",
    "django-cors-headers>=4.1.0",
    "djangorestframework-simplejwt>=5.2.2",
    "setuptools>=68.1.2",
]
requires-python = ">=3.10"
license = {text = "MIT"}
